feat: Improve fan command defaults and PWM conversion
Default fan actions to all headers and add a short profile option.
Use wider arithmetic when converting percentages to PWM values.

+++ src/utils.rs
@@ -1,5 +1,6 @@
 #[inline]
+/// Converts a percentage (0-100) to a PWM value (0-255)
 pub const fn pct(p: u8) -> u8 {
     let p = if p > 100 { 100 } else { p };
-    ((p * 255 + 99) / 100) as u8
+    ((p as u32 * 255 + 99) / 100) as u8
 }
